[MINOR] fix QuickstartUtils move
